Add support for reporting YANG model features in the capabilities
This change allows a user to specify any supported features in a YANG
model. Netconf parses any files of the type .info looking for lines
which specify a YANG model name and a comma separated list of features.
