+++ models/netconf-features.info
@@ -1,0 +1,1 @@
+example ether,fast
